[refactor] : 포인트 서비스 레이어에서 Money VO 사용으로 변경
 - PointService 인터페이스의 메서드 파라미터를 int에서 Money로 변경
 - PointEvent DTO의 totalPrice 필드를 Money 타입으로 변경
 - 서비스 레이어 전반에서 타입 안전한 금액 처리를 위한 Money VO 활용
