+++ src/main/java/shop/sajotuna/order/point/service/PointHistoryWriter.java
@@ -2,6 +2,7 @@ package shop.sajotuna.order.point.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import shop.sajotuna.order.common.domain.Money;
 import shop.sajotuna.order.point.domain.PointHistory;
 import shop.sajotuna.order.point.repository.PointHistoryRepository;
 
@@ -13,11 +14,11 @@ public class PointHistoryWriter {
 
     private final PointHistoryRepository pointHistoryRepository;
 
-    public void savePointEarnHistory(Long userId, int pointAmount, String description) {
+    public void savePointEarnHistory(Long userId, Money pointAmount, String description) {
         PointHistory pointHistory = pointHistoryRepository.save(PointHistory.createEarnHistory(userId, pointAmount, description));
     }
 
-    public void savePointRedeemHistory(Long userId, int pointAmount) {
+    public void savePointRedeemHistory(Long userId, Money pointAmount) {
         PointHistory pointHistory = pointHistoryRepository.save(PointHistory.createRedeemHistory(userId, pointAmount, REDEEM_MESSAGE));
     }
 }
