+++ src/main/java/shop/sajotuna/order/point/service/PointQueueService.java
@@ -1,6 +1,6 @@
 package shop.sajotuna.order.point.service;
 
-import shop.sajotuna.order.point.controller.request.PointEvent;
+import shop.sajotuna.order.point.service.dto.event.PointEvent;
 
 public interface PointQueueService {
     void sendEarnPointsMessage(PointEvent event);
