+++ src/main/java/shop/sajotuna/order/point/service/PointService.java
@@ -1,5 +1,6 @@
 package shop.sajotuna.order.point.service;
 
+import shop.sajotuna.order.common.domain.Money;
 import shop.sajotuna.order.point.controller.response.PointHistoryResponse;
 
 import java.util.List;
@@ -8,5 +9,5 @@ public interface PointService {
 
     List<PointHistoryResponse> getPointsByUserId(Long userId);
 
-    PointHistoryResponse redeemPoints(Long userId, int pointAmount);
+    PointHistoryResponse redeemPoints(Long userId, Money pointAmount);
 }
